Rename model package to 'rest'.
The reasoning is that in the future we might want to support different persistence
strategies, including local storage and web sockets. Those packages can now have natural
names and different requirements for what a model must contain. E.g. a local storage Model
does not need to have a RootURL method. I think that rest is also more inline with what the
package actually does currently. Existing javascript libraries tend to wrap this functionality
into a module called 'http' or 'resource'. 'http' is already taken by the standard library,
and I think 'resource' is a little ambiguous compared to 'rest'.

Additionally since rest.Model is still an interface, we allow a single type to satisfy multiple
Model interfaces and therefore multiple persistence strategies. This is useful for e.g., saving
a model in local storage first and periodically syncing to a server, or falling back to using
a REST API if a browser does not support web sockets.