Remove errors.go.
The only code in this file was an error that was returned when we couldn't find the element
for a particular view in the DOM by using a QuerySelector for its id. Since we are no longer
going to be using view ids, this error message doesn't belong anymore. It is probably more
appropriate to put named errors in the relevant sub-package anyways, and right now I can't
think of any named error type that would be shared accross packages. If we do end up making
shared error types, we can create the errors.go file again.